+++ include/zenoh-pico/collections/list.h
@@ -34,7 +34,7 @@ typedef struct _z_l_t
 _z_list_t *_z_list_of(void *x);
 
 size_t _z_list_len(const _z_list_t *xs);
-int _z_list_is_empty(const _z_list_t *xs);
+uint8_t _z_list_is_empty(const _z_list_t *xs);
 
 void *_z_list_head(const _z_list_t *xs);
 _z_list_t *_z_list_tail(const _z_list_t *xs);
@@ -59,7 +59,7 @@ void _z_list_free(_z_list_t **xs, z_element_free_f f_f);
     {                                                                                                \
         return _z_list_len(l);                                                                       \
     }                                                                                                \
-    static inline int name##_list_is_empty(const name##_list_t *l)                                   \
+    static inline uint8_t name##_list_is_empty(const name##_list_t *l)                               \
     {                                                                                                \
         return _z_list_is_empty(l);                                                                  \
     }                                                                                                \
